feat/read a amiibo's nickname from the VirtualAmiiboFile
This feature adds a way to change the Amiibo's nickname inside Smash and other places where it's used, so it’s not always 'Ryujinx.' However, it was not added a GUI or create the Cabinet applet that would allow users to change this. So you will have to go to system/amiibo and find your amiibo id to change it.