fix/window sizing when 'Show Title Bar' is enabled
Fixes a bug that causes the main window to not size properly when the TitleBar is enabled (i.e.: when the TitleBar and MenuStrip are separate entities). Corrects the size for main window startup and when a user clicks a 'View > Window Size > *Resolution Here*' MenuStripItem. Prior to this fix if a user selects 720p/1080p and 'Show Title Bar' is enabled, the window would be sized smaller than intended and display black bars on the sides of the render area